+++ src/frontend/apps/impress/src/core/AppProvider.tsx
@@ -3,10 +3,10 @@ import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { useEffect } from 'react';
 
 import { useCunninghamTheme } from '@/cunningham';
+import { Auth } from '@/features/auth';
 import '@/i18n/initI18n';
 import { useResponsiveStore } from '@/stores/';
 
-import { Auth } from './auth/';
 import { ConfigProvider } from './config/';
 
 /**
