--- src/frontend/apps/impress/src/core/auth/index.ts
@@ -1,4 +1,0 @@
-export * from './api/types';
-export * from './Auth';
-export * from './ButtonLogin';
-export * from './useAuthStore';
