+++ src/frontend/apps/impress/src/core/index.ts
@@ -1,3 +1,2 @@
 export * from './AppProvider';
-export * from './auth';
 export * from './config';
