+++ src/frontend/apps/impress/src/features/docs/doc-share/api/useCreateDocAccess.tsx
@@ -1,7 +1,7 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 
 import { APIError, errorCauses, fetchAPI } from '@/api';
-import { User } from '@/core/auth';
+import { User } from '@/features/auth';
 import {
   Access,
   Doc,
