+++ src/frontend/apps/impress/src/features/docs/doc-share/api/useCreateDocInvitation.tsx
@@ -1,7 +1,7 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 
 import { APIError, errorCauses, fetchAPI } from '@/api';
-import { User } from '@/core/auth';
+import { User } from '@/features/auth';
 import { Doc, Role } from '@/features/docs/doc-management';
 import { Invitation, OptionType } from '@/features/docs/doc-share/types';
 import { ContentLanguage } from '@/i18n/types';
