+++ src/frontend/apps/impress/src/features/docs/doc-share/api/useUsers.tsx
@@ -1,7 +1,7 @@
 import { UseQueryOptions, useQuery } from '@tanstack/react-query';
 
 import { APIError, APIList, errorCauses, fetchAPI } from '@/api';
-import { User } from '@/core/auth';
+import { User } from '@/features/auth';
 import { Doc } from '@/features/docs/doc-management';
 
 export type UsersParams = {
