+++ src/frontend/apps/impress/src/features/docs/doc-share/types.tsx
@@ -1,4 +1,4 @@
-import { User } from '@/core/auth';
+import { User } from '@/features/auth';
 import { Role } from '@/features/docs';
 
 export interface Invitation {
